+++ package.json
@@ -2,11 +2,19 @@
   "name": "@map-colonies/flat-earth",
   "version": "1.0.0",
   "description": "Package for geospatial and map tiles calculations",
-  "main": "build/src/index.js",
-  "types": "dist/types.d.ts",
-  "files": [
-    "dist"
-  ],
+  "types": "dist/types/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "exports": {
+    "import": {
+      "types": "./dist/types/index.d.ts",
+      "default": "./dist/esm/index.js"
+    },
+    "require": {
+      "types": "./dist/types/index.d.ts",
+      "default": "./dist/cjs/index.js"
+    }
+  },
   "license": "ISC",
   "keywords": [
     "geospatial",
@@ -14,8 +22,11 @@
     "tiles"
   ],
   "scripts": {
-    "compile": "tsc",
-    "prepublish": "npm run compile",
+    "build:types": "tsc -p tsconfig.types.json",
+    "build:cjs": "tsc -p tsconfig.cjs.json && node postcjs.js",
+    "build:esm": "tsc -p tsconfig.esm.json && node postesm.js",
+    "build": "rimraf ./dist && npm run build:types && npm run build:cjs && npm run build:esm",
+    "prepublish": "npm run build",
     "test": "npm run test:node && npm run test:web",
     "test:node": "jest -c tests/jest.config.js",
     "test:web": "jest -c tests/jest-web.config.js",
@@ -46,6 +57,7 @@
     "jest-extended": "^4.0.2",
     "prettier": "3.2.4",
     "pretty-quick": "^4.0.0",
+    "rimraf": "^6.0.1",
     "ts-jest": "^29.1.2",
     "typedoc": "^0.25.12",
     "typescript": "~5.4.2"
