+++ postcjs.js
@@ -1,0 +1,1 @@
+require('fs').writeFileSync('./dist/cjs/package.json', '{"type":"commonjs"}', 'utf-8');
