+++ postesm.js
@@ -1,0 +1,1 @@
+require('fs').writeFileSync('./dist/esm/package.json', '{"type":"module"}', 'utf-8');
