+++ src/tiles/tiles.ts
@@ -1,9 +1,8 @@
 import type { BBox, Position } from 'geojson';
-import type { ReverseIntersectionPolicy } from '../types';
 import type { ArrayElement } from '../utils/types';
 import { validatePositionByTileMatrix } from '../validations/validations';
 import type { TileMatrixSet } from './tileMatrixSet';
-import type { TileIndex, TileMatrix, TileMatrixId } from './types';
+import type { ReverseIntersectionPolicy, TileIndex, TileMatrix, TileMatrixId } from './types';
 
 export function tileMatrixToBBox<T extends TileMatrixSet>(
   tileMatrix: ArrayElement<T['tileMatrices']>,
