+++ src/tiles/types.ts
@@ -2,6 +2,7 @@ import { ArrayElement } from '../utils/types';
 import { CORNER_OF_ORIGIN_CODE } from './constants';
 
 export type Comparison = 'equal' | 'closest' | 'lower' | 'higher';
+export type ReverseIntersectionPolicy = 'none' | 'both' | 'row' | 'col';
 
 /**
  * Tile index
