+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "target": "ES2021",
+    "target": "es2021",
     "module": "commonjs",
-    "declaration": true,
-    "outDir": "dist",
+    // "declaration": true,
+    // "outDir": "dist",
     "strict": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
